+++ src/lib.rs
@@ -244,7 +244,23 @@ fn eval(ast: Type, env: &Rc<Env>) -> Ret {
 
                         Type::Symbol(symbol) if symbol == "try*" => {
                             // (try* A (catch* B C))
-                            nargs_check("try*", 2, list.len() - 1)?;
+                            nargs_check("try*", 1, list.len() - 1)
+                                .or_else(|_| error::nargs_check("try*", 2, list.len() - 1))?;
+
+                            // try*
+                            let try_body = list[1].to_owned();
+                            let try_result = eval(try_body, env);
+
+                            if try_result.is_ok() || list.len() == 2 {
+                                return try_result;
+                            }
+
+                            // catch*
+                            let exception = match try_result.unwrap_err() {
+                                Exception::Builtin(s) => Type::String(s),
+                                Exception::Custom(t) => t,
+                            };
+
                             let catch = match list[2].to_owned() {
                                 Type::List(seq) | Type::Vector(seq) => seq,
                                 _ => {
@@ -255,31 +271,17 @@ fn eval(ast: Type, env: &Rc<Env>) -> Ret {
                             };
                             nargs_check("catch*", 2, catch.len() - 1)?;
 
-                            let try_body = list[1].to_owned();
                             let catch_var = match &catch[1] {
                                 Type::Symbol(s) => s,
-                                _ => return Err(Exception::type_error(
-                                    "First catch* argument must be a symbol",
-                                )),
+                                _ => {
+                                    return Err(Exception::type_error(
+                                        "First catch* argument must be a symbol",
+                                    ))
+                                }
                             };
                             let catch_body = catch[2].to_owned();
 
-                            // try*
-                            let try_result = eval(try_body, env);
-
-                            if try_result.is_ok() {
-                                return try_result;
-                            }
-
-                            let exception = try_result.unwrap_err();
-                            let exception = Type::String(exception.to_string());
-
-                            // catch*
-                            let exc_env = Env::new(
-                                Some(env.clone()),
-                                &[catch_var],
-                                &[exception],
-                            );
+                            let exc_env = Env::new(Some(env.clone()), &[catch_var], &[exception]);
 
                             Ok(eval(catch_body, &Rc::new(exc_env))?)
                         }